+++ react/components/UserAddress/AddressInfo.js
@@ -1,23 +1,56 @@
 import React, { Fragment } from 'react'
 import { ExtensionPoint, useChildBlock } from 'vtex.render-runtime'
+import { useCssHandles } from 'vtex.css-handles'
 import { pathOr } from 'ramda'
 import { injectIntl, FormattedMessage } from 'react-intl'
 import { IconLocationMarker } from 'vtex.store-icons'
 
-const AddressInfo = ({ inverted, inline, orderForm, intl }) => {
+const CSS_HANDLES = [
+  'addressInfoIconContainer',
+  'addressInfoTextContainer',
+  'addressInfoPrefixContainer',
+  'addressInfoAddressContainer',
+  'addressInfoDivider',
+  'addressInfoModalContainer',
+]
+
+const AddressInfo = ({
+  inverted,
+  inline,
+  orderForm,
+  intl,
+  showStreet,
+  showCityAndState,
+  showPostalCode,
+  showPrefix,
+}) => {
   const { shippingData } = orderForm
   const hasModal = !!useChildBlock({ id: 'modal' })
+  const handles = useCssHandles(CSS_HANDLES)
 
   if (!shippingData || !shippingData.address) return
 
-  const { street, number, complement, addressType } = shippingData.address
+  const {
+    street,
+    number,
+    complement,
+    addressType,
+    city,
+    state,
+    postalCode,
+  } = shippingData.address
 
-  const displayStreet = number ? `${street}, ${number}` : street
+  let displayStreet = number ? `${street}, ${number}` : street
 
-  const displayAddress =
-    complement && complement !== ''
-      ? `${displayStreet} - ${complement}`
-      : `${displayStreet}`
+  if (complement) displayStreet = `${displayStreet} - ${complement}`
+
+  const displayCityAndState = `${city}, ${state}`
+
+  const displayAddress = `${showStreet ? displayStreet : ''}${
+    showStreet && (showCityAndState || showPostalCode) ? ', ' : ''
+  }${showCityAndState ? displayCityAndState : ''}${
+    showCityAndState && showPostalCode ? ', ' : ''
+  }${showPostalCode ? postalCode : ''}`
 
   const isPickup = addressType === 'pickup'
   const friendlyName = pathOr(
@@ -28,43 +61,53 @@ const AddressInfo = ({ inverted, inline, orderForm, intl }) => {
 
   return (
     <div className={`flex ${inline ? 'items-end' : 'items-center flex-auto'}`}>
-      <div className="flex flex-auto">
+      <div className="flex flex-auto items-center">
         <div
-          className={`mr3 flex items-center ${
+          className={`${
+            handles.addressInfoIconContainer
+          } mr3 flex items-center ${
             inverted ? 'c-on-base--inverted' : 'c-muted-2'
           }`}
         >
           <IconLocationMarker size={27} viewBox="0 0 21 27" />
         </div>
-        <div className="flex flex-auto flex-column">
-          <div
-            className={`t-small ${
-              inverted ? 'c-on-base--inverted' : 'c-muted-2'
-            }`}
-          >
-            {isPickup ? (
-              <FormattedMessage
-                id="store/user-address.pickup"
-                values={{ name: friendlyName }}
-              />
-            ) : (
-              <FormattedMessage id="store/user-address.order" />
-            )}
+        <div
+          className={`${handles.addressInfoTextContainer} flex flex-auto flex-column`}
+        >
+          {showPrefix && (
+            <div
+              className={`${handles.addressInfoPrefixContainer} t-small ${
+                inverted ? 'c-on-base--inverted' : 'c-muted-2'
+              }`}
+            >
+              {isPickup ? (
+                <FormattedMessage
+                  id="store/user-address.pickup"
+                  values={{ name: friendlyName }}
+                />
+              ) : (
+                <FormattedMessage id="store/user-address.order" />
+              )}
+            </div>
+          )}
+          <div className={`${handles.addressInfoAddressContainer} truncate`}>
+            {displayAddress}
           </div>
-          <div className="truncate">{displayAddress}</div>
         </div>
       </div>
       {hasModal && (
         <Fragment>
           <div
-            className={`bl bw1 mh4 ${inline ? 'nb2' : ''} ${
-              inverted ? 'b--on-base--inverted' : 'b--muted-5'
-            }`}
+            className={`${handles.addressInfoDivider} bl bw1 mh4 ${
+              inline ? 'nb2' : ''
+            } ${inverted ? 'b--on-base--inverted' : 'b--muted-5'}`}
             style={{
               height: '1.5rem',
             }}
           />
-          <div className="flex items-center">
+          <div
+            className={`${handles.addressInfoModalContainer} flex items-center`}
+          >
             <ExtensionPoint
               id="modal"
               centered